15-utilities/MameyBarcode/requirements.txt
Flask==2.0.1
python-barcode==0.13.1
pdf417gen==0.6.0
qrcode==7.3
Pillow==8.3.1
Flask-HTTPAuth==4.4.0
Flask-Limiter==1.4
